+++ packages/rollup-dev-tools/src/tools/transpiler/index.ts
@@ -7,4 +7,4 @@ export * from './opcode-whitelist'
 export * from './static-memory-opcodes'
 export * from './transpiler'
 export * from './util'
-export * from './log-search-generation'
+export * from './binary-search-tree'
